requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cloudinary-storage
django-cors-headers
django-filter
Pillow
python-decouple
stripe
drf-spectacular
social-auth-app-django==5.4.0
django-allauth==0.60.1
django-email-verification==0.3.3
django-ratelimit==4.1.0
django-argon2
django-axes
django-csp